migration-kahoots: add demos
We're going to be running an interactive session with a customer that
involves running through some demos, followed by some Kahoots questions.

These are largely derived from demos we already have, with some changes
to fit them into the time constraints and tailor the demos to the
questions and points we want to cover.